# crm_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MDIS CRM Backend API",
        "title": "MDIS CRM API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/analytics/kpi": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Review activity (calls) and outcome (success deals) per admissions manager.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Manager performance KPI",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/crm_backend_internal_model.ManagerKPI"
                            }
                        }
                    }
                }
            }
        },
...
        "/integrations/telegram/webhook": {
            "post": {
                "description": "Inbound endpoint for Telegram Bot updates. Maps chats to Leads.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "summary": "Telegram Webhook",
                "parameters": [
                    {
                        "description": "Telegram Update",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/crm_backend_internal_model.TelegramWebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/interactions": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Log a call, SMS, or message against a lead or deal.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interactions"
                ],
                "summary": "Record a new interaction",
                "parameters": [
                    {
                        "description": "Interaction Data",
                        "name": "interaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/crm_backend_internal_model.CreateInteractionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/crm_backend_internal_model.Interaction"
                        }
                    }
                }
            }
        },
...
        "/leads": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of the 100 most recent leads.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "List leads",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/crm_backend_internal_model.Lead"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new lead from a form submission, automatically linking a contact and opening a deal.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "Create a new lead",
                "parameters": [
                    {
                        "description": "Create Lead Request",
                        "name": "lead",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/crm_backend_internal_model.CreateLeadRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/crm_backend_internal_model.Lead"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "crm_backend_internal_model.CreateInteractionRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deal_id": {
                    "type": "integer"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "lead_id": {
                    "type": "integer"
                },
                "type": {
                    "description": "\"call\", \"sms\", \"email\", \"messenger\", \"manual_note\"",
                    "type": "string"
                }
            }
        },
        "crm_backend_internal_model.CreateLeadRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "program_id": {
                    "type": "integer"
                },
                "utm_campaign": {
                    "type": "string"
                },
                "utm_medium": {
                    "type": "string"
                },
                "utm_source": {
                    "type": "string"
                }
            }
        },
...
        "crm_backend_internal_model.Interaction": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deal_id": {
                    "type": "integer"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lead_id": {
                    "type": "integer"
                },
                "type": {
                    "description": "\"call\", \"sms\", \"email\", \"messenger\", \"manual_note\"",
                    "type": "string"
                }
            }
        },
        "crm_backend_internal_model.Lead": {
            "type": "object",
            "properties": {
                "assignee_id": {
                    "type": "integer"
                },
                "contact_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "program_id": {
                    "type": "integer"
                },
                "source_id": {
                    "type": "integer"
                },
                "status_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "utm_campaign": {
                    "type": "string"
                },
                "utm_medium": {
                    "type": "string"
                },
                "utm_source": {
                    "type": "string"
                }
            }
        },
...
        "crm_backend_internal_model.ManagerKPI": {
            "type": "object",
            "properties": {
                "calls_count": {
                    "type": "integer"
                },
                "closed_deals": {
                    "description": "count deals moved to 'Zachislenie'",
                    "type": "integer"
                },
                "manager_id": {
                    "type": "integer"
                },
                "manager_name": {
                    "type": "string"
                },
                "revenue_added": {
                    "description": "Placeholder if payments are tracked",
                    "type": "number"
                }
            }
        },
        "crm_backend_internal_model.TelegramWebhookRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "object",
                    "properties": {
                        "chat": {
                            "type": "object",
                            "properties": {
                                "first_name": {
                                    "type": "string"
                                },
                                "id": {
                                    "type": "integer"
                                },
                                "type": {
                                    "type": "string"
                                },
                                "username": {
                                    "type": "string"
                                }
                            }
                        },
                        "date": {
                            "type": "integer"
                        },
                        "from": {
                            "type": "object",
                            "properties": {
                                "first_name": {
                                    "type": "string"
                                },
                                "id": {
                                    "type": "integer"
                                },
                                "is_bot": {
                                    "type": "boolean"
                                },
                                "language_code": {
                                    "type": "string"
                                },
                                "username": {
                                    "type": "string"
                                }
                            }
                        },
                        "message_id": {
                            "type": "integer"
                        },
                        "text": {
                            "type": "string"
                        }
                    }
                },
                "update_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}